PIM-4252: move the validator step to the new component
