+++ src/Pim/Bundle/ConnectorBundle/Resources/config/batch_jobs.yml
@@ -7,9 +7,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_product_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_product_import.import.title
                     services:
@@ -28,9 +28,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_attribute_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_attribute_import.import.title
                     services:
@@ -43,9 +43,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_attribute_option_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_attribute_option_import.import.title
                     services:
@@ -58,9 +58,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_variant_group_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_variant_group_import.import.title
                     services:
@@ -75,9 +75,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_category_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_category_import.import.title
                     services:
@@ -92,9 +92,9 @@ connector:
             steps:
                 validation:
                     title: pim_base_connector.jobs.csv_group_import.validation.title
-                    class: %pim_base_connector.step.validator.class%
+                    class: %pim_connector.step.validator.class%
                     services:
-                        charsetValidator: pim_base_connector.validator.step.charset_validator
+                        charsetValidator: pim_connector.validator.item.charset_validator
                 import:
                     title:         pim_connector.jobs.csv_group_import.import.title
                     services:
