+++ src/Pim/Bundle/ConnectorBundle/Resources/config/items.yml
@@ -1,0 +1,6 @@
+parameters:
+    pim_connector.validator.item.charset_validator.class: Pim\Component\Connector\Item\CharsetValidator
+
+services:
+    pim_connector.validator.item.charset_validator:
+        class: %pim_connector.validator.item.charset_validator.class%
